ceval: fix detection of Python function calls
We want to record profile (and trace) events of Python function calls
after the FUNC_HEADER is executed, but before any subsequent opcodes.
The strategy is to check that ts->pc (the last pc) points to FUNC_HEADER
and pc points to the next opcode. This strategy is kind of fragile and
frequent source of bugs, but it mostly works...

One bug was that handling the eval breaker in between FUNC_HEADER and
the subsequent opcode overwrote ts->pc with the to-be-executed pc. This
modifies vm_eval_breaker to restore the ts->pc with its previous value.